Handle incomplete files
dlvideos.py:
If done.json does not exist, check each file in _videos using ffmpeg. If
the file is incomplete (i.e., the regex '^Truncating' matches ffmpeg's
stderr on one or more lines), it is moved to the directory '_incomplete'
and the download is re-run.
Regardless, after downloading each video, check it using ffmpeg, and if
the file is incomplete as per above, don't add the URL to done.json,
making sure it's re-downloaded next time dlvideos.py is run.